Use non-forked stylis for minification
